+++ public/nav.php
@@ -38,6 +38,7 @@
           <li><a href="/advanced.php?submitted=true&amp;p=1&amp;is_dupe=true">All duplicate files</a></li>
         </ul>
       </li>
+			<li><a href="/sunburst.php?path=/">File Tree</a></li>
     </ul>
     <form method="get" action="/simple.php" class="navbar-form navbar-left" role="search">
       <div class="form-group">